Flip CoreOS to Flatcar
