+++ flatcar/amiregistry/reliable_http.go
