+++ flatcar/amiregistry/reliable_http_test.go
